# S5.01_API_REST/fitbit_api/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Fitbit API Documentation",
        "description": "API para gestión de disciplinas, comunidades y usuarios",
        "contact": {
            "email": "admin@fitbit.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://my-default-host.com",
            "description": "API Server"
        }
    ],
    "paths": {
...
        "/api/v1/disciplines": {
            "get": {
                "tags": [
                    "Disciplines"
                ],
                "summary": "Obtener lista de disciplinas",
                "description": "Devuelve todas las disciplinas disponibles (acceso público)",
                "operationId": "getDisciplines",
                "responses": {
                    "200": {
                        "description": "Lista de disciplinas obtenida exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Discipline"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Disciplines"
                ],
                "summary": "Crear nueva disciplina",
                "description": "Crea una nueva disciplina (solo administradores)",
                "operationId": "storeDiscipline",
                "requestBody": {
                    "description": "Datos de la disciplina",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "maxLength": 150,
                                        "example": "Boxing"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Arte marcial de combate con puños",
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Disciplina creada exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Discipline created successfully"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/Discipline"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "No autorizado"
                    },
                    "403": {
                        "description": "Sin permisos de administrador"
                    },
                    "422": {
                        "description": "Error de validación"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
        "/api/v1/disciplines/{id}": {
            "get": {
                "tags": [
                    "Disciplines"
                ],
                "summary": "Obtener disciplina específica",
                "description": "Devuelve una disciplina específica por ID (acceso público)",
                "operationId": "showDiscipline",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la disciplina",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Disciplina encontrada",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/Discipline"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Disciplina no encontrada"
                    }
                }
            },
            "put": {
                "tags": [
                    "Disciplines"
                ],
                "summary": "Actualizar disciplina",
                "description": "Actualiza una disciplina existente (solo administradores)",
                "operationId": "updateDiscipline",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la disciplina",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "requestBody": {
                    "description": "Datos actualizados de la disciplina",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "maxLength": 150,
                                        "example": "Boxing Updated"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Descripción actualizada",
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Disciplina actualizada exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Discipline updated successfully"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/Discipline"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "No autorizado"
                    },
                    "403": {
                        "description": "Sin permisos"
                    },
                    "404": {
                        "description": "Disciplina no encontrada"
                    },
                    "422": {
                        "description": "Error de validación"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Disciplines"
                ],
                "summary": "Eliminar disciplina",
                "description": "Elimina una disciplina existente (solo administradores)",
                "operationId": "destroyDiscipline",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la disciplina",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Disciplina eliminada exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Discipline deleted successfully"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "No autorizado"
                    },
                    "403": {
                        "description": "Sin permisos"
                    },
                    "404": {
                        "description": "Disciplina no encontrada"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Discipline": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "Boxing"
                    },
                    "description": {
                        "type": "string",
                        "example": "Arte marcial de combate"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearer_token": {
                "type": "http",
                "description": "Enter token in format (Bearer <token>)",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            },
            "type": "http",
            "description": "HTTP authentication",
            "name": "Authorization",
            "in": "header",
            "scheme": "bearer"
        }
    },
...
    "security": [
        {
            "bearer_token": []
        }
    ]
}